Add files via upload
The dataset consists of Turbofan Engine Degradation Simulation data made up of 12 text files created by NASA.
It contains sensor data generated by 100 turbofan engines of the same model. Each turbofan engine starts operating normally and ends with a failure.
Each engine has a lifecycle, and at the end of this period, the engine fails.

On this dataset, the aim is to perform the following tasks:

Estimating the Remaining Useful Life (RUL) of the engine (Regression Analysis),

Predicting whether the engine is in a normal state or approaching failure (Binary Classification),

Predicting whether the engine is in a normal, warning, or alarm state before failure (Multi-class Classification).